Backup rotate script added and set to be called after backup is done
Created /sbin/e-smith/backup-config-rotate
The script will check that only 15 backups are keept ( value is
hardcoded )

backup-config-rotate will be called after the do_backup() will be called
.
This way, after a new backup is created the old backup is deleted

The /sbin/e-smith/backup-config-rotate will have to also be added to the
sudoers file so it can be called

Also minor fixes to the backup php file
Failed to properly add styling to the qtip. It seems that the Jquery UI
takes precedence or the default is used.
Inline styles are ignored.