UI changes and added new functionality
Restore.php - Has been removed because it is replaced with upload from
the BackupConfig page

Backup.php - Has received new UI features and tweaks.
- Added a new form to upload backup archives directly from the backup
page.
- Added progress bar when restoring a backup. This will prompt the user
to wait 90 seconds until the changes are applied and the backup has
finished restoring.

- UI tweaks - added MD5 check.
The md5 for each archive file is checked with the one in the .md5 file.
If the checksums do not match a warning is shown. Also if the md5 file
is not found for a specific backup, a warning is shown.

- The Backups are sorted with the latest one shown first in the table.