BackupConfig module Beta
Beta version of the BackupConfig module.
Changes:
Added script that moves the uploaded backup archive from the UI to the
backups folder
Added the used scripts to the sudoers
Modified the info div to keep the list of actions performed. Now it
behaves somewhat like a log file

No known issues were found at this time.

If there are no issues reported with this module it will become final V
1.0